#123 refactor: move mocks to their own package
