+++ aljebra/src/test/java/com/aljebra/field/mock/FieldCalls.java
@@ -21,7 +21,7 @@
  * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.aljebra.field;
+package com.aljebra.field.mock;
 
 import com.aljebra.scalar.Scalar;
 import java.util.ArrayList;
