+++ aljebra/src/test/java/com/aljebra/field/mock/MkOrderedRandom.java
@@ -21,8 +21,9 @@
  * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.aljebra.field;
+package com.aljebra.field.mock;
 
+import com.aljebra.field.OrderedRandom;
 import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
