+++ aljebra/src/test/java/com/aljebra/field/mock/SpyField.java
@@ -21,8 +21,12 @@
  * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.aljebra.field;
+package com.aljebra.field.mock;
 
+import com.aljebra.field.Field;
+import com.aljebra.field.FieldAddition;
+import com.aljebra.field.FieldMultiplication;
+import com.aljebra.field.MetricSpaceField;
 import com.aljebra.scalar.Scalar;
 
 /**
