+++ aljebra/src/test/java/com/aljebra/field/mock/SpyMetricSpace.java
@@ -21,8 +21,10 @@
  * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.aljebra.field;
+package com.aljebra.field.mock;
 
+import com.aljebra.field.AbstractField;
+import com.aljebra.field.MetricSpaceField;
 import com.aljebra.metric.InnerProduct;
 import com.aljebra.scalar.Scalar;
 import java.util.Arrays;
