+++ aljebra/src/test/java/com/aljebra/scalar/AddInverseTest.java
@@ -23,9 +23,9 @@
  */
 package com.aljebra.scalar;
 
-import com.aljebra.field.MkAddition;
-import com.aljebra.field.MkField;
-import com.aljebra.field.SpyField;
+import com.aljebra.field.mock.MkAddition;
+import com.aljebra.field.mock.MkField;
+import com.aljebra.field.mock.SpyField;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Test;
