+++ aljebra/src/test/java/com/aljebra/scalar/condition/TernaryTest.java
@@ -23,7 +23,7 @@
  */
 package com.aljebra.scalar.condition;
 
-import com.aljebra.field.SpyField;
+import com.aljebra.field.mock.SpyField;
 import com.aljebra.scalar.Scalar;
 import java.util.List;
 import java.util.Optional;
